Fix abort warning and add explicit abort method
Previously we were using two different atomic sources to discover when a response has been dropped, but only emitting a warning for an unconsumed response body if _one_ of the sources indicated a drop. This was causing inconsistent results of when the warning would be emitted. Fix this by having just one source of truth for when the response is dropped.

In addition, add `ResponseExt::abort`, which allows you to abort a response stream explicitly without any warning being emitted if you intentionally want abort behavior, instead of simply dropping.

See #270.